# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: RedApp HTTP API
  version: unreleased
  summary: Public catalog, distribution paths and administration API of RedApp.
  license:
    name: MIT
    identifier: MIT
  description: |
    This document is the single source of truth for every HTTP route RedApp serves.
    The Go server (`internal/httpserver`) and the Vue frontend are implemented from it;
    contract tests validate real responses against it. Design notes are in `docs/dev/api.md`.

    ## Route families

    | Prefix | Audience | Format |
    | --- | --- | --- |
    | `/api/` | anonymous public API | JSON (one HTML document) |
    | `/admin/api/` | administrator API, session + CSRF | JSON (ZIP export, multipart uploads) |
    | `/<vendor>/<app>/<file_path>` | distribution: installers, release metadata, artifacts, hosted and HTTP cache files | binary/text |
    | `/`, `/all`, `/<vendor>`, `/<vendor>/<app>`, `/admin/...` | SPA pages | HTML |
    | `/assets/` | frontend build output, uploaded icons, preset images | static |
    | `/health/` | probes | JSON |

    ## Common rules

    - **Methods.** GET reads and is safe to retry; PUT replaces a whole settings document;
      PATCH applies a sparse change; POST creates or runs an action; DELETE removes.
      Every route registered for GET also answers HEAD (Go `http.ServeMux` semantics) with the same
      headers and no body; HEAD is listed explicitly only where it matters (files, icons).
...
      return `400 INVALID_QUERY`.
...

    ## Errors
...
servers:
...
paths:
...
  /admin/api/apps:
    get:
      operationId: listApps
      tags: [directory]
      summary: List applications, optionally of one vendor, with sorting.
      description: |
        `q` matches application ID, `vendor/app` key, both names and tags (tags: case-folded NFC substring).
        Sorting happens over all matching applications before paging; ties are broken by key and missing
        values always sort last regardless of `order`.
        `version`/`updated` use the newest locally known version of the current source epoch (protocol order);
        `downloads` uses the lifetime `counters.download_success` (only `http-cache`, `codex`, `claude-code`).
      parameters:
        - name: vendor
          in: query
          description: Restrict to one vendor (deleted vendors allowed).
          schema:
            $ref: '#/components/schemas/VendorId'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/DirectoryStateQuery'
        - name: sort
          in: query
          description: Sort key.
          schema:
            type: string
            enum: [name, version, updated, downloads]
            default: name
        - name: order
          in: query
          description: Sort direction.
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: lang
          in: query
          description: Name language used by `sort=name` (default `en`).
          schema:
            $ref: '#/components/schemas/Language'
        - $ref: '#/components/parameters/Page'
        - name: limit
          in: query
          description: Page size.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      security:
        - sessionCookie: []
      x-error-codes: [VENDOR_NOT_FOUND]
      responses:
        '200':
          description: One page of applications.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      operationId: createApp
      tags: [directory]
      summary: Create a custom application under an existing vendor.
      description: |
        The provider, vendor and ID are fixed after creation. Provider-specific fields:
        `codex`/`claude-code` accept `base_url` (default: provider default) and `cache_ttl_seconds` (channel TTL 1..86400, default 60);
        `http-cache` requires `base_urls` (1..16) and accepts `source_strategy` and `cache_ttl_seconds` (0..86400, default 300);
        `info`/`hosted` accept none of them. Proxy starts as `inherit`. Release applications get the
        default retention (`enabled: false, keep_latest: 3`) and prewarm (`enabled: false`) settings.
      security:
        - sessionCookie: []
          csrfToken: []
      x-max-body-bytes: 65536
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppCreate'
            examples:
              httpCache:
                value:
                  vendor: example
                  id: tools
                  provider: http-cache
                  name: {en: Tools, zh-CN: 工具}
                  description: {en: Internal tool mirror, zh-CN: 内部工具镜像}
                  icon: ''
                  enabled: false
                  categories: []
                  tags: [mirror]
                  base_urls: [https://files.example.internal/tools, https://mirror.example.internal/tools]
                  source_strategy: ordered
                  cache_ttl_seconds: 300
      x-error-codes: [VENDOR_NOT_FOUND, VALIDATION_FAILED, ALREADY_EXISTS, ENTITY_DELETED]
      responses:
        '201':
          description: Created application.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              $ref: '#/components/headers/Location'
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /admin/api/apps/{vendor}/{app}/prewarm/options:
    parameters:
      - $ref: '#/components/parameters/Vendor'
      - $ref: '#/components/parameters/App'
    get:
      operationId: getPrewarmOptions
      tags: [prewarm]
      summary: Read what a prewarm task of this application can target.
      description: Available for `codex`, `claude-code` (kind `release`) and `http-cache` (kind `http_cache`).
      security:
        - sessionCookie: []
      x-error-codes: [APPLICATION_NOT_FOUND, CAPABILITY_UNSUPPORTED]
      responses:
        '200':
          description: Prewarm options.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrewarmOptions'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /admin/api/apps/{vendor}/{app}/prewarm/jobs:
    parameters:
      - $ref: '#/components/parameters/Vendor'
      - $ref: '#/components/parameters/App'
    post:
      operationId: startPrewarm
      tags: [prewarm]
      summary: Start a one-time prewarm task.
      description: |
        Only one prewarm task runs in the whole service; there is no queue (`409 PREWARM_BUSY` while another runs).
        `request_id` is a client-generated idempotency key: repeating the same request returns the existing job with `200`;
        reusing it with a different body, or after the job expired, returns `409 PREWARM_REQUEST_CONFLICT`.
        Finished jobs are kept for 24 hours; a restart marks a running job `interrupted`. Cancelling never cancels downloads
        that public clients also wait for. Prewarm never counts as public downloads. A disabled application cannot be
        prewarmed (`409 APPLICATION_DISABLED`).
      x-idempotency: '`request_id` (32 hex characters) per application, valid while the job is retained.'
      x-polling: The admin UI polls `getPrewarmJob` every 1.5 seconds while `state` is `running`.
      security:
        - sessionCookie: []
          csrfToken: []
      x-max-body-bytes: 8388608
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrewarmStartRequest'
            examples:
              release:
                value:
                  request_id: 9f2c4e0a1b3d5f7e9a1c3e5f7b9d1f3a
                  target: latest
                  platforms: [linux-x64, darwin-arm64]
                  limits: {max_files: 10000, max_depth: 16, max_download_bytes: 10737418240, max_duration_seconds: 3600}
              httpCache:
                value:
                  request_id: 1a2b3c4d5e6f708192a3b4c5d6e7f809
                  paths: [/releases/1.2.3/tool.zip]
                  indexes: [/releases/]
                  match: {type: glob, pattern: '/**/*.zip'}
                  limits: {max_files: 500, max_depth: 4, max_download_bytes: 21474836480, max_duration_seconds: 7200}
      x-error-codes: [APPLICATION_NOT_FOUND, CAPABILITY_UNSUPPORTED, VALIDATION_FAILED, PREWARM_BUSY, PREWARM_REQUEST_CONFLICT, SOURCE_CHANGED, ENTITY_DELETED, APPLICATION_DISABLED]
      responses:
        '200':
          description: Existing job for a repeated `request_id`.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrewarmJob'
        '201':
          description: New job started.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrewarmJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /admin/api/apps/{vendor}/{app}/prewarm/jobs/{job_id}:
    parameters:
      - $ref: '#/components/parameters/Vendor'
      - $ref: '#/components/parameters/App'
      - $ref: '#/components/parameters/JobId'
    get:
      operationId: getPrewarmJob
      tags: [prewarm]
      summary: Read a prewarm job.
      security:
        - sessionCookie: []
      x-error-codes: [APPLICATION_NOT_FOUND, CAPABILITY_UNSUPPORTED, JOB_NOT_FOUND]
      responses:
        '200':
          description: The job.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrewarmJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /admin/api/apps/{vendor}/{app}/prewarm/jobs/{job_id}/items:
    parameters:
      - $ref: '#/components/parameters/Vendor'
      - $ref: '#/components/parameters/App'
      - $ref: '#/components/parameters/JobId'
    get:
      operationId: listPrewarmItems
      tags: [prewarm]
      summary: List the planned files of a prewarm job with their results.
      description: In planning order.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit25'
      security:
        - sessionCookie: []
      x-error-codes: [APPLICATION_NOT_FOUND, CAPABILITY_UNSUPPORTED, JOB_NOT_FOUND]
      responses:
        '200':
          description: One page of items.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrewarmItemPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /admin/api/apps/{vendor}/{app}/prewarm/jobs/{job_id}/cancel:
    parameters:
      - $ref: '#/components/parameters/Vendor'
      - $ref: '#/components/parameters/App'
      - $ref: '#/components/parameters/JobId'
    post:
      operationId: cancelPrewarmJob
      tags: [prewarm]
      summary: Request cancellation of a running prewarm job.
      description: Asynchronous; the job moves to `cancelled` shortly. Cancelling a finished job is a no-op that returns it.
      security:
        - sessionCookie: []
          csrfToken: []
      x-error-codes: [APPLICATION_NOT_FOUND, CAPABILITY_UNSUPPORTED, JOB_NOT_FOUND]
      responses:
        '202':
          description: Cancellation requested; current job state.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrewarmJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /admin/api/apps/{vendor}/{app}/prewarm/jobs/{job_id}/retry:
    parameters:
      - $ref: '#/components/parameters/Vendor'
      - $ref: '#/components/parameters/App'
      - $ref: '#/components/parameters/JobId'
    post:
      operationId: retryPrewarmJob
      tags: [prewarm]
      summary: Start a new job that retries the unsuccessful items of a finished job.
      description: |
        Successful items are skipped; a release job is pinned to the version the original job resolved.
        Same idempotency and busy rules as `startPrewarm`.
      x-idempotency: '`request_id` as in `startPrewarm`.'
      security:
        - sessionCookie: []
          csrfToken: []
      x-max-body-bytes: 8192
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrewarmRetryRequest'
      x-error-codes: [APPLICATION_NOT_FOUND, CAPABILITY_UNSUPPORTED, VALIDATION_FAILED, JOB_NOT_FOUND, OPERATION_IN_PROGRESS, PREWARM_BUSY, PREWARM_REQUEST_CONFLICT, SOURCE_CHANGED, ENTITY_DELETED, APPLICATION_DISABLED]
      responses:
        '200':
          description: Existing job for a repeated `request_id`.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrewarmJob'
        '201':
          description: New job started.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrewarmJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  # ---------------------------------------------------------------- hosted
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: redapp_session
      description: Administrator session cookie set by `createSession`. Sessions live in memory (lost on restart).
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: The 64 hex character `csrf_token` of the session. Required on every unsafe admin method.
  parameters:
    Vendor:
      name: vendor
      in: path
      required: true
      description: Vendor ID.
      schema:
        $ref: '#/components/schemas/VendorId'
    App:
      name: app
      in: path
      required: true
      description: Application ID within the vendor.
      schema:
        $ref: '#/components/schemas/Slug'
...
    JobId:
      name: job_id
      in: path
      required: true
      description: Prewarm job ID.
      schema:
        $ref: '#/components/schemas/Uid'
...
    Page:
      name: page
      in: query
      description: 1-based page number.
      schema:
        type: integer
        minimum: 1
        maximum: 1000000000
        default: 1
...
    SearchQuery:
      name: q
      in: query
      description: Search text, trimmed, at most 128 characters.
      schema:
        type: string
        maxLength: 128
    DirectoryStateQuery:
      name: state
      in: query
      description: Lifecycle filter.
      schema:
        $ref: '#/components/schemas/DirectoryState'
...
  headers:
    ETag:
      description: Quoted revision of the returned resource.
      schema:
        type: string
        pattern: '^"[1-9][0-9]{0,17}"$'
    XRequestId:
      description: Request ID generated by the server for this request; also written to logs.
      schema:
        type: string
        pattern: '^[0-9a-f]{16}$'
    Location:
      description: Admin API path of the created resource.
      schema:
        type: string
  responses:
...
    BadRequest:
      description: The request is malformed or invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: {code: VALIDATION_FAILED, message: 'cache_ttl_seconds must be between 0 and 86400', request_id: 4f2a9c0d1e3b5a7c, retryable: false}
    Unauthorized:
      description: No valid session, or wrong password.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: {code: AUTH_REQUIRED, message: Sign in required, request_id: 4f2a9c0d1e3b5a7c, retryable: false}
    Forbidden:
      description: Origin or CSRF check failed.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: {code: CSRF_REJECTED, message: CSRF validation failed, request_id: 4f2a9c0d1e3b5a7c, retryable: false}
    NotFound:
      description: The addressed object does not exist or is not visible.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: {code: APPLICATION_NOT_FOUND, message: Application not found, request_id: 4f2a9c0d1e3b5a7c, retryable: false}
...
    Conflict:
      description: The request conflicts with the current state.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: {code: REVISION_CONFLICT, message: Changed by someone else; reload before saving, request_id: 4f2a9c0d1e3b5a7c, retryable: false}
    PayloadTooLarge:
      description: Request body, upload or artifact exceeds its limit.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnsupportedMediaType:
      description: Wrong request `Content-Type`.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Unexpected server error (recovered panic).
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ServiceUnavailable:
      description: Local storage, capacity or a dependency is temporarily unavailable.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Retry-After:
          description: Seconds until a retry may succeed, when known.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    VendorId:
      description: Vendor ID; `admin`, `api`, `assets`, `health` and `all` are reserved.
      allOf:
        - $ref: '#/components/schemas/Slug'
        - not:
            enum: [admin, api, assets, health, all]
...
    Language:
      type: string
      enum: [en, zh-CN]
...
    App:
      type: object
      additionalProperties: false
      required: [uid, id, key, vendor_uid, vendor_id, name, description, icon, provider, categories, tags, enabled, builtin_template, source_epoch, revision, deleted_at]
      description: |
        Administrator view. Upstream fields depend on the provider: `base_url` (`codex`, `claude-code`),
        `base_urls` and `source_strategy` (`http-cache`), `cache_ttl_seconds` (all three); absent otherwise.
      properties:
        uid:
          $ref: '#/components/schemas/Uid'
        id:
          $ref: '#/components/schemas/Slug'
        key:
          $ref: '#/components/schemas/AppKey'
        vendor_uid:
          $ref: '#/components/schemas/Uid'
        vendor_id:
          $ref: '#/components/schemas/VendorId'
        name:
          $ref: '#/components/schemas/LocalizedText'
        description:
          $ref: '#/components/schemas/LocalizedText'
        icon:
          $ref: '#/components/schemas/IconPath'
        provider:
          $ref: '#/components/schemas/ProviderKey'
        base_url:
          type: string
        base_urls:
          type: array
          minItems: 1
          maxItems: 16
          items:
            type: string
        source_strategy:
          $ref: '#/components/schemas/SourceStrategy'
        cache_ttl_seconds:
          type: integer
          minimum: 0
          maximum: 86400
          description: Channel TTL for release providers (≥ 1); default freshness TTL for `http-cache`.
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Slug'
        tags:
          type: array
          items:
            type: string
        enabled:
          type: boolean
        builtin_template:
          type: boolean
          description: Built-in (cannot be deleted).
        source_epoch:
          type: integer
          format: int64
          minimum: 0
          description: Current source epoch (`0` for providers without upstream).
        revision:
          $ref: '#/components/schemas/Revision'
        deleted_at:
          type: [string, 'null']
          format: date-time
...
    AppPage:
      type: object
      additionalProperties: false
      required: [items, page, limit, total, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AppListItem'
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          format: int64
          minimum: 0
        total_pages:
          type: integer
          minimum: 1
    AppCreate:
      type: object
      additionalProperties: false
      required: [vendor, id, provider, name, enabled]
      properties:
        vendor:
          $ref: '#/components/schemas/VendorId'
        id:
          $ref: '#/components/schemas/Slug'
        provider:
          $ref: '#/components/schemas/ProviderKey'
        name:
          $ref: '#/components/schemas/Name'
        description:
          $ref: '#/components/schemas/Description'
        icon:
          $ref: '#/components/schemas/IconPath'
        enabled:
          type: boolean
        categories:
          type: array
          maxItems: 32
          items:
            $ref: '#/components/schemas/Slug'
        tags:
          $ref: '#/components/schemas/Tags'
        base_url:
          $ref: '#/components/schemas/HttpUrl'
        base_urls:
          type: array
          minItems: 1
          maxItems: 16
          uniqueItems: true
          items:
            $ref: '#/components/schemas/HttpUrl'
        source_strategy:
          $ref: '#/components/schemas/SourceStrategy'
        cache_ttl_seconds:
          type: integer
          minimum: 0
          maximum: 86400
...
    PrewarmOptions:
      type: object
      additionalProperties: false
      required: [kind, channels, platforms, default_limits]
      properties:
        kind:
          type: string
          enum: [release, http_cache]
        channels:
          type: array
          description: Release channels (empty for `http_cache`).
          items:
            type: string
        platforms:
          type: array
          description: Release platforms (empty for `http_cache`).
          items:
            $ref: '#/components/schemas/Platform'
        default_limits:
          $ref: '#/components/schemas/PrewarmLimits'
    PrewarmStartRequest:
      description: Use the variant that matches `PrewarmOptions.kind`.
      oneOf:
        - $ref: '#/components/schemas/ReleasePrewarmRequest'
        - $ref: '#/components/schemas/HttpCachePrewarmRequest'
...